carry fn to proving.rs
